Move cli parsing to its own file
