+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nethermindeth/warp",
-  "version": "2.4.3",
+  "version": "2.4.4",
   "description": "Solidity to Cairo Transpiler",
   "main": "build/export.js",
   "author": "Nethermind",
