+++ src/starknetCli.ts
@@ -8,7 +8,7 @@ import {
   IDeployAccountProps,
   IOptionalDebugInfo,
   IDeclareOptions,
-} from './index';
+} from './cli';
 import { CLIError, logError } from './utils/errors';
 import { callClassHashScript } from './utils/utils';
 import { encodeInputs } from './transcode/encode';
