+++ src/utils/astPrinter.ts
@@ -8,7 +8,7 @@ import {
   TypeNameType,
   TypeNode,
 } from 'solc-typed-ast';
-import { PrintOptions } from '..';
+import { PrintOptions } from '../cli';
 import { CairoFunctionDefinition, FunctionStubKind } from '../ast/cairoNodes';
 import { cyan, underline } from './formatting';
 import { extractProperty } from './utils';
