+++ src/utils/utils.ts
@@ -66,7 +66,7 @@ import {
 } from './nodeTemplates';
 import { isDynamicArray, isDynamicCallDataArray, safeGetNodeType } from './nodeTypeProcessing';
 import { Class } from './typeConstructs';
-import { TranspilationOptions } from '..';
+import { TranspilationOptions } from '../cli';
 
 const uint128 = BigInt('0x100000000000000000000000000000000');
